// src/models/FeatureIdent.ts
export enum FeatureIdent {
  DEV_MODE="DEV_MODE",
  BOOKMARKS="BOOKMARKS",
  STATS="STATS",
  GROUP_BY_DOMAIN = "GROUP_BY_DOMAIN",
  RSS = "RSS",
  SPACES = "SPACES",
  TOP10 = "TOP10",
  OPEN_TABS = "OPEN_TABS",
  WINDOWS = "WINDOWS",
  BACKUP = "BACKUP",
  NOTES = "NOTES",
  TAGS = "TAGS",
  STANDALONE_APP = "STANDALONE_APP",
  NOTIFICATIONS = "NOTIFICATIONS",
  COLOR_TAGS = "COLOR_TAGS",
  ADVANCED_TAB_MANAGEMENT = "ADVANCED_TAB_MANAGEMENT",
  TAB_GROUPS = "TAB_GROUPS",
  TABS_AS_TREE = "TABS_AS_TREE",
  WINDOWS_MANAGEMENT = "WINDOWS_MANAGEMENT",
  NEWEST_TABS="NEWEST_TABS",
  TABSETS_SHARING = "TABSETS_SHARING"
}

export enum FeatureType {
  RECOMMENDED = "RECOMMENDED",
  OPTIONAL = "OPTIONAL",
  EXPERIMENTAL = "EXPERIMENTAL",
  PLANNED = "PLANNED",
  DISABLED = "DISABLED",
  INTERNAL = "INTERNAL"
}